re-organizing sql logic
