+++ crates/scouter_sql/src/sql/mod.rs
@@ -1,4 +1,5 @@
 pub mod postgres;
 pub mod query;
 pub mod schema;
+pub mod traits;
 pub mod utils;
