+++ crates/scouter_sql/src/sql/postgres.rs
@@ -1,44 +1,33 @@
 use crate::sql::query::Queries;
 use crate::sql::schema::{
-    AlertWrapper, BinnedCustomMetricWrapper, Entity, FeatureBinProportionResultWrapper,
-    FeatureBinProportionWrapper, ObservabilityResult, SpcFeatureResult, TaskRequest,
-    UpdateAlertResult, User,
+    AlertWrapper, Entity, ObservabilityResult, TaskRequest, UpdateAlertResult, User,
 };
+use crate::sql::traits::{CustomMetricSqlLogic, PsiSqlLogic, SpcSqlLogic};
 use crate::sql::utils::pg_rows_to_server_records;
 use chrono::{DateTime, Utc};
 use cron::Schedule;
 use scouter_contracts::{
-    DriftAlertRequest, DriftRequest, GetProfileRequest, ObservabilityMetricRequest,
-    ProfileStatusRequest, ServiceInfo, UpdateAlertStatus,
+    DriftAlertRequest, GetProfileRequest, ObservabilityMetricRequest, ProfileStatusRequest,
+    ServiceInfo, UpdateAlertStatus,
 };
-use scouter_dataframe::parquet::{
-    dataframe_to_custom_drift_metrics, dataframe_to_psi_drift_features,
-    dataframe_to_spc_drift_features, ParquetDataFrame,
-};
+
 use scouter_error::{ScouterError, SqlError, UtilError};
 use scouter_settings::{DatabaseSettings, ObjectStorageSettings};
-use scouter_types::psi::FeatureBinProportionResult;
 use scouter_types::DriftType;
 use scouter_types::{
-    alert::Alert,
-    custom::BinnedCustomMetrics,
-    psi::FeatureBinProportions,
-    spc::{SpcDriftFeature, SpcDriftFeatures},
-    CustomMetricServerRecord, DriftProfile, ObservabilityMetrics, PsiServerRecord, RecordType,
-    ServerRecords, SpcServerRecord, TimeInterval, ToDriftRecords,
+    alert::Alert, DriftProfile, ObservabilityMetrics, RecordType, ServerRecords, TimeInterval,
+    ToDriftRecords,
 };
 use serde_json::Value;
 use sqlx::{
     postgres::{PgPoolOptions, PgQueryResult},
     Pool, Postgres, Row, Transaction,
 };
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
 use std::result::Result::Ok;
 use std::str::FromStr;
 use std::sync::Arc;
 use tracing::{debug, error, info, instrument};
-
-use super::utils::split_custom_interval;
 
 // TODO: Explore refactoring and breaking this out into multiple client types (i.e., spc, psi, etc.)
 // Postgres client is one of the lowest-level abstractions so it may not be worth it, as it could make server logic annoying. Worth exploring though.
@@ -50,6 +39,10 @@ pub struct PostgresClient {
     pub retention_period: i64,
     pub storage_settings: ObjectStorageSettings,
 }
+
+impl SpcSqlLogic for PostgresClient {}
+impl CustomMetricSqlLogic for PostgresClient {}
+impl PsiSqlLogic for PostgresClient {}
 
 impl PostgresClient {
     /// Create a new PostgresClient
@@ -220,50 +213,6 @@ impl PostgresClient {
             Ok(result) => Ok(result),
             Err(e) => Err(SqlError::traced_query_error(e.to_string())),
         }
-    }
-
-    /// Inserts a drift record into the database
-    ///
-    /// # Arguments
-    ///
-    /// * `record` - A drift record to insert into the database
-    /// * `table_name` - The name of the table to insert the record into
-    ///
-    pub async fn insert_spc_drift_record(
-        &self,
-        record: &SpcServerRecord,
-    ) -> Result<PgQueryResult, SqlError> {
-        let query = Queries::InsertDriftRecord.get_query();
-
-        sqlx::query(&query.sql)
-            .bind(record.created_at)
-            .bind(&record.name)
-            .bind(&record.space)
-            .bind(&record.version)
-            .bind(&record.feature)
-            .bind(record.value)
-            .execute(&self.pool)
-            .await
-            .map_err(SqlError::traced_query_error)
-    }
-
-    pub async fn insert_bin_counts(
-        &self,
-        record: &PsiServerRecord,
-    ) -> Result<PgQueryResult, SqlError> {
-        let query = Queries::InsertBinCounts.get_query();
-
-        sqlx::query(&query.sql)
-            .bind(record.created_at)
-            .bind(&record.name)
-            .bind(&record.space)
-            .bind(&record.version)
-            .bind(&record.feature)
-            .bind(record.bin_id as i64)
-            .bind(record.bin_count as i64)
-            .execute(&self.pool)
-            .await
-            .map_err(SqlError::traced_query_error)
     }
 
     // Inserts a drift record into the database
@@ -445,76 +394,6 @@ impl PostgresClient {
         }
     }
 
-    // Queries the database for all features under a service
-    // Private method that'll be used to run drift retrieval in parallel
-    pub async fn get_spc_features(
-        &self,
-        service_info: &ServiceInfo,
-    ) -> Result<Vec<String>, SqlError> {
-        let query = Queries::GetSpcFeatures.get_query();
-
-        sqlx::query(&query.sql)
-            .bind(&service_info.name)
-            .bind(&service_info.space)
-            .bind(&service_info.version)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(SqlError::traced_get_features_error)
-            .map(|result| {
-                result
-                    .iter()
-                    .map(|row| row.get("feature"))
-                    .collect::<Vec<String>>()
-            })
-    }
-
-    /// Get SPC drift records
-    ///
-    /// # Arguments
-    ///
-    /// * `service_info` - The service to get drift records for
-    /// * `limit_datetime` - The limit datetime to get drift records for
-    /// * `features_to_monitor` - The features to monitor
-    pub async fn get_spc_drift_records(
-        &self,
-        service_info: &ServiceInfo,
-        limit_datetime: &DateTime<Utc>,
-        features_to_monitor: &[String],
-    ) -> Result<SpcDriftFeatures, SqlError> {
-        let mut features = self.get_spc_features(service_info).await?;
-
-        if !features_to_monitor.is_empty() {
-            features.retain(|feature| features_to_monitor.contains(feature));
-        }
-
-        let query = Queries::GetSpcFeatureValues.get_query();
-
-        let records: Vec<SpcFeatureResult> = sqlx::query_as(&query.sql)
-            .bind(limit_datetime)
-            .bind(&service_info.name)
-            .bind(&service_info.space)
-            .bind(&service_info.version)
-            .bind(features)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(SqlError::traced_query_error)?;
-
-        let feature_drift = records
-            .into_iter()
-            .map(|record| {
-                let feature = SpcDriftFeature {
-                    created_at: record.created_at,
-                    values: record.values,
-                };
-                (record.feature.clone(), feature)
-            })
-            .collect::<BTreeMap<String, SpcDriftFeature>>();
-
-        Ok(SpcDriftFeatures {
-            features: feature_drift,
-        })
-    }
-
     pub async fn get_binned_observability_metrics(
         &self,
         params: &ObservabilityMetricRequest,
@@ -536,190 +415,6 @@ impl PostgresClient {
                 .await;
 
         observability_metrics.map_err(SqlError::traced_query_error)
-    }
-
-    // Queries the database for drift records based on a time window and aggregation
-    //
-    // # Arguments
-    //
-    // * `name` - The name of the service to query drift records for
-    // * `space` - The name of the space to query drift records for
-    // * `feature` - The name of the feature to query drift records for
-    // * `aggregation` - The aggregation to use for the query
-    // * `time_interval` - The time window to query drift records for
-    //
-    // # Returns
-    //
-    // * A vector of drift records
-    pub async fn get_binned_spc_drift_records(
-        &self,
-        params: &DriftRequest,
-    ) -> Result<SpcDriftFeatures, SqlError> {
-        let minutes = params.time_interval.to_minutes();
-        let bin = params.time_interval.to_minutes() as f64 / params.max_data_points as f64;
-
-        let query = Queries::GetBinnedSpcFeatureValues.get_query();
-
-        let records: Vec<SpcFeatureResult> = sqlx::query_as(&query.sql)
-            .bind(bin)
-            .bind(minutes)
-            .bind(&params.name)
-            .bind(&params.space)
-            .bind(&params.version)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(SqlError::traced_query_error)?;
-
-        let feature_drift = records
-            .into_iter()
-            .map(|record| {
-                let feature = SpcDriftFeature {
-                    created_at: record.created_at,
-                    values: record.values,
-                };
-                (record.feature.clone(), feature)
-            })
-            .collect::<BTreeMap<String, SpcDriftFeature>>();
-
-        Ok(SpcDriftFeatures {
-            features: feature_drift,
-        })
-    }
-
-    // Queries the database for drift records based on a time window and aggregation
-    //
-    // # Arguments
-    //
-    // * `name` - The name of the service to query drift records for
-    // * `params` - The drift request parameters
-    // # Returns
-    //
-    // * A vector of drift records
-    pub async fn get_binned_psi_drift_records(
-        &self,
-        params: &DriftRequest,
-    ) -> Result<Vec<FeatureBinProportionResult>, SqlError> {
-        let mut results = Vec::new();
-        if params.has_custom_interval() {
-            let mut feature_map: BTreeMap<String, FeatureBinProportionResult> = BTreeMap::new();
-            let interval = params.clone().custom_interval.unwrap();
-            let timestamps =
-                split_custom_interval(interval.start, interval.end, &self.retention_period)?;
-
-            if let Some((archive_begin, archive_end)) = timestamps.archived_range {
-                let path = format!("{}/{}/{}/psi", params.space, params.name, params.version);
-                let archived_minutes = timestamps.archived_minutes.unwrap() as f64;
-                let bin = archived_minutes / params.max_data_points as f64;
-
-                let archived_df = ParquetDataFrame::new(&self.storage_settings, &RecordType::Psi)?
-                    .get_binned_metrics(
-                        &path,
-                        &bin,
-                        &archive_begin,
-                        &archive_end,
-                        &params.space,
-                        &params.name,
-                        &params.version,
-                    )
-                    .await?;
-
-                let archived_results = dataframe_to_psi_drift_features(archived_df)
-                    .await
-                    .map_err(SqlError::traced_failed_to_convert_dataframe_error)?;
-
-                for result in archived_results {
-                    feature_map
-                        .entry(result.feature.clone())
-                        .and_modify(|existing| {
-                            existing.created_at.extend(result.created_at);
-                            existing.bin_proportions.extend(result.bin_proportions);
-                            // Merge overall proportions
-                            for (k, v) in result.overall_proportions {
-                                existing.overall_proportions.insert(k, v);
-                            }
-                        })
-                        .or_insert(result);
-                }
-            }
-
-            if let Some(minutes) = timestamps.current_minutes {
-                let query = Queries::GetBinnedPsiFeatureBins.get_query();
-                let bin = minutes as f64 / params.max_data_points as f64;
-                let current_results: Vec<FeatureBinProportionResult> = sqlx::query_as(&query.sql)
-                    .bind(bin)
-                    .bind(minutes)
-                    .bind(&params.name)
-                    .bind(&params.space)
-                    .bind(&params.version)
-                    .fetch_all(&self.pool)
-                    .await
-                    .map_err(SqlError::traced_query_error)?
-                    .into_iter()
-                    .map(|wrapper: FeatureBinProportionResultWrapper| wrapper.0)
-                    .collect();
-
-                results.extend(current_results);
-            }
-        } else {
-            // get features
-
-            let minutes = params.time_interval.to_minutes();
-            let bin = params.time_interval.to_minutes() as f64 / params.max_data_points as f64;
-            let query = Queries::GetBinnedPsiFeatureBins.get_query();
-            let binned: Vec<FeatureBinProportionResult> = sqlx::query_as(&query.sql)
-                .bind(bin)
-                .bind(minutes)
-                .bind(&params.name)
-                .bind(&params.space)
-                .bind(&params.version)
-                .fetch_all(&self.pool)
-                .await
-                .map_err(SqlError::traced_query_error)?
-                .into_iter()
-                .map(|wrapper: FeatureBinProportionResultWrapper| wrapper.0)
-                .collect();
-
-            results.extend(binned);
-        }
-
-        results.sort_by(|a, b| a.timestamp.cmp(&b.timestamp));
-
-        Ok(results)
-    }
-
-    // Queries the database for drift records based on a time window and aggregation
-    //
-    // # Arguments
-    //
-    // * `name` - The name of the service to query drift records for
-    // * `params` - The drift request parameters
-    // # Returns
-    //
-    // * A vector of drift records
-    pub async fn get_binned_custom_drift_records(
-        &self,
-        params: &DriftRequest,
-    ) -> Result<BinnedCustomMetrics, SqlError> {
-        // get features
-
-        let minutes = params.time_interval.to_minutes();
-        let bin = params.time_interval.to_minutes() as f64 / params.max_data_points as f64;
-
-        let query = Queries::GetBinnedCustomMetricValues.get_query();
-
-        let records: Vec<BinnedCustomMetricWrapper> = sqlx::query_as(&query.sql)
-            .bind(bin)
-            .bind(minutes)
-            .bind(&params.name)
-            .bind(&params.space)
-            .bind(&params.version)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(SqlError::traced_query_error)?;
-
-        Ok(BinnedCustomMetrics::from_vec(
-            records.into_iter().map(|wrapper| wrapper.0).collect(),
-        ))
     }
 
     pub async fn update_drift_profile_status(
@@ -744,81 +439,6 @@ impl PostgresClient {
                 error!("Failed to update drift profile status: {:?}", e);
                 Err(SqlError::traced_update_drift_profile_error(e))
             }
-        }
-    }
-
-    pub async fn get_feature_bin_proportions(
-        &self,
-        service_info: &ServiceInfo,
-        limit_datetime: &DateTime<Utc>,
-        features_to_monitor: &[String],
-    ) -> Result<FeatureBinProportions, SqlError> {
-        let query = Queries::GetFeatureBinProportions.get_query();
-
-        let binned: Vec<FeatureBinProportionWrapper> = sqlx::query_as(&query.sql)
-            .bind(&service_info.name)
-            .bind(&service_info.space)
-            .bind(&service_info.version)
-            .bind(limit_datetime)
-            .bind(features_to_monitor)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(SqlError::traced_get_bin_proportions_error)?;
-
-        let binned: FeatureBinProportions = binned.into_iter().map(|wrapper| wrapper.0).collect();
-
-        Ok(binned)
-    }
-
-    pub async fn get_custom_metric_values(
-        &self,
-        service_info: &ServiceInfo,
-        limit_datetime: &DateTime<Utc>,
-        metrics: &[String],
-    ) -> Result<HashMap<String, f64>, SqlError> {
-        let query = Queries::GetCustomMetricValues.get_query();
-
-        let records = sqlx::query(&query.sql)
-            .bind(&service_info.name)
-            .bind(&service_info.space)
-            .bind(&service_info.version)
-            .bind(limit_datetime)
-            .bind(metrics)
-            .fetch_all(&self.pool)
-            .await
-            .map_err(SqlError::traced_get_custom_metrics_error)?;
-
-        let metric_map = records
-            .into_iter()
-            .map(|row| {
-                let metric = row.get("metric");
-                let value = row.get("value");
-                (metric, value)
-            })
-            .collect();
-
-        Ok(metric_map)
-    }
-
-    pub async fn insert_custom_metric_value(
-        &self,
-        record: &CustomMetricServerRecord,
-    ) -> Result<PgQueryResult, SqlError> {
-        let query = Queries::InsertCustomMetricValues.get_query();
-
-        let query_result = sqlx::query(&query.sql)
-            .bind(record.created_at)
-            .bind(&record.name)
-            .bind(&record.space)
-            .bind(&record.version)
-            .bind(&record.metric)
-            .bind(record.value)
-            .execute(&self.pool)
-            .await;
-
-        match query_result {
-            Ok(result) => Ok(result),
-            Err(e) => Err(SqlError::traced_insert_custom_metrics_error(e)),
         }
     }
 
@@ -1041,9 +661,12 @@ impl MessageHandler {
                         debug!("SPC record count: {:?}", records.len());
                         let records = records.to_spc_drift_records()?;
                         for record in records.iter() {
-                            let _ = client.insert_spc_drift_record(record).await.map_err(|e| {
-                                error!("Failed to insert drift record: {:?}", e);
-                            });
+                            let _ = client
+                                .insert_spc_drift_record(&client.pool, record)
+                                .await
+                                .map_err(|e| {
+                                    error!("Failed to insert drift record: {:?}", e);
+                                });
                         }
                     }
                     RecordType::Observability => {
@@ -1062,9 +685,12 @@ impl MessageHandler {
                         debug!("PSI record count: {:?}", records.len());
                         let records = records.to_psi_drift_records()?;
                         for record in records.iter() {
-                            let _ = client.insert_bin_counts(record).await.map_err(|e| {
-                                error!("Failed to insert bin count record: {:?}", e);
-                            });
+                            let _ = client
+                                .insert_bin_counts(&client.pool, record)
+                                .await
+                                .map_err(|e| {
+                                    error!("Failed to insert bin count record: {:?}", e);
+                                });
                         }
                     }
                     RecordType::Custom => {
@@ -1072,7 +698,7 @@ impl MessageHandler {
                         let records = records.to_custom_metric_drift_records()?;
                         for record in records.iter() {
                             let _ = client
-                                .insert_custom_metric_value(record)
+                                .insert_custom_metric_value(&client.pool, record)
                                 .await
                                 .map_err(|e| {
                                     error!("Failed to insert bin count record: {:?}", e);
@@ -1092,8 +718,9 @@ mod tests {
 
     use super::*;
     use rand::Rng;
-    use scouter_types::{spc::SpcDriftProfile, DriftType};
-
+    use scouter_contracts::DriftRequest;
+    use scouter_types::spc::SpcDriftProfile;
+    use scouter_types::*;
     const SPACE: &str = "space";
     const NAME: &str = "name";
     const VERSION: &str = "1.0.0";
@@ -1225,7 +852,10 @@ mod tests {
             value: 1.0,
         };
 
-        let result = client.insert_spc_drift_record(&record).await.unwrap();
+        let result = client
+            .insert_spc_drift_record(&client.pool, &record)
+            .await
+            .unwrap();
 
         assert_eq!(result.rows_affected(), 1);
     }
@@ -1244,7 +874,10 @@ mod tests {
             bin_count: 1,
         };
 
-        let result = client.insert_bin_counts(&record).await.unwrap();
+        let result = client
+            .insert_bin_counts(&client.pool, &record)
+            .await
+            .unwrap();
 
         assert_eq!(result.rows_affected(), 1);
     }
@@ -1332,7 +965,10 @@ mod tests {
                     value: j as f64,
                 };
 
-                let result = client.insert_spc_drift_record(&record).await.unwrap();
+                let result = client
+                    .insert_spc_drift_record(&client.pool, &record)
+                    .await
+                    .unwrap();
                 assert_eq!(result.rows_affected(), 1);
             }
         }
@@ -1343,26 +979,32 @@ mod tests {
             version: VERSION.to_string(),
         };
 
-        let features = client.get_spc_features(&service_info).await.unwrap();
+        let features = client
+            .get_spc_features(&client.pool, &service_info)
+            .await
+            .unwrap();
         assert_eq!(features.len(), 10);
 
         let records = client
-            .get_spc_drift_records(&service_info, &timestamp, &features)
+            .get_spc_drift_records(&client.pool, &service_info, &timestamp, &features)
             .await
             .unwrap();
 
         assert_eq!(records.features.len(), 10);
 
         let binned_records = client
-            .get_binned_spc_drift_records(&DriftRequest {
-                space: SPACE.to_string(),
-                name: NAME.to_string(),
-                version: VERSION.to_string(),
-                time_interval: TimeInterval::FiveMinutes,
-                max_data_points: 10,
-                drift_type: DriftType::Spc,
-                custom_interval: None,
-            })
+            .get_binned_spc_drift_records(
+                &client.pool,
+                &DriftRequest {
+                    space: SPACE.to_string(),
+                    name: NAME.to_string(),
+                    version: VERSION.to_string(),
+                    time_interval: TimeInterval::FiveMinutes,
+                    max_data_points: 10,
+                    drift_type: DriftType::Spc,
+                    custom_interval: None,
+                },
+            )
             .await
             .unwrap();
 
@@ -1388,13 +1030,17 @@ mod tests {
                         bin_count: rand::rng().random_range(0..10),
                     };
 
-                    client.insert_bin_counts(&record).await.unwrap();
+                    client
+                        .insert_bin_counts(&client.pool, &record)
+                        .await
+                        .unwrap();
                 }
             }
         }
 
         let binned_records = client
             .get_feature_bin_proportions(
+                &client.pool,
                 &ServiceInfo {
                     space: SPACE.to_string(),
                     name: NAME.to_string(),
@@ -1417,15 +1063,20 @@ mod tests {
         assert!(*bin_proportion > 0.1 && *bin_proportion < 0.2);
 
         let binned_records = client
-            .get_binned_psi_drift_records(&DriftRequest {
-                space: SPACE.to_string(),
-                name: NAME.to_string(),
-                version: VERSION.to_string(),
-                time_interval: TimeInterval::OneHour,
-                max_data_points: 1000,
-                drift_type: DriftType::Psi,
-                custom_interval: None,
-            })
+            .get_binned_psi_drift_records(
+                &client.pool,
+                &DriftRequest {
+                    space: SPACE.to_string(),
+                    name: NAME.to_string(),
+                    version: VERSION.to_string(),
+                    time_interval: TimeInterval::OneHour,
+                    max_data_points: 1000,
+                    drift_type: DriftType::Psi,
+                    custom_interval: None,
+                },
+                &client.retention_period,
+                &client.storage_settings,
+            )
             .await
             .unwrap();
         //
@@ -1449,7 +1100,10 @@ mod tests {
                     value: rand::rng().random_range(0..10) as f64,
                 };
 
-                let result = client.insert_custom_metric_value(&record).await.unwrap();
+                let result = client
+                    .insert_custom_metric_value(&client.pool, &record)
+                    .await
+                    .unwrap();
                 assert_eq!(result.rows_affected(), 1);
             }
         }
@@ -1464,11 +1118,15 @@ mod tests {
             value: rand::rng().random_range(0..10) as f64,
         };
 
-        let result = client.insert_custom_metric_value(&record).await.unwrap();
+        let result = client
+            .insert_custom_metric_value(&client.pool, &record)
+            .await
+            .unwrap();
         assert_eq!(result.rows_affected(), 1);
 
         let metrics = client
             .get_custom_metric_values(
+                &client.pool,
                 &ServiceInfo {
                     space: SPACE.to_string(),
                     name: NAME.to_string(),
@@ -1483,15 +1141,18 @@ mod tests {
         assert_eq!(metrics.len(), 1);
 
         let binned_records = client
-            .get_binned_custom_drift_records(&DriftRequest {
-                space: SPACE.to_string(),
-                name: NAME.to_string(),
-                version: VERSION.to_string(),
-                time_interval: TimeInterval::OneHour,
-                max_data_points: 1000,
-                drift_type: DriftType::Custom,
-                custom_interval: None,
-            })
+            .get_binned_custom_drift_records(
+                &client.pool,
+                &DriftRequest {
+                    space: SPACE.to_string(),
+                    name: NAME.to_string(),
+                    version: VERSION.to_string(),
+                    time_interval: TimeInterval::OneHour,
+                    max_data_points: 1000,
+                    drift_type: DriftType::Custom,
+                    custom_interval: None,
+                },
+            )
             .await
             .unwrap();
         //
